+++ tests/testthat/test-search-functions.R
@@ -115,12 +115,12 @@ test_that("Check link generation - Posit Community", {
 
   expect_identical(
     search_posit("toad"),
-    "https://community.posit.co/search?q=toad"
+    "https://forum.posit.co/search?q=toad"
   )
 
   expect_identical(
     search_posit("toad", rlang = FALSE),
-    "https://community.posit.co/search?q=toad"
+    "https://forum.posit.co/search?q=toad"
   )
 
 })
@@ -138,6 +138,36 @@ test_that("Check link generation - twitter", {
   )
 
 })
+
+
+test_that("Check link generation - Mastodon", {
+
+  expect_identical(
+    search_mastodon("toad"),
+    "https://mastodon.social/search?q=toad %23rstats"
+  )
+
+  expect_identical(
+    search_mastodon("toad", rlang = FALSE),
+    "https://mastodon.social/search?q=toad"
+  )
+
+})
+
+test_that("Check link generation - BlueSky", {
+
+  expect_identical(
+    search_bluesky("toad"),
+    "https://bsky.app/search?q=toad %23rstats"
+  )
+
+  expect_identical(
+    search_bluesky("toad", rlang = FALSE),
+    "https://bsky.app/search?q=toad"
+  )
+
+})
+
 
 
 test_that("Check link generation - stackoverflow", {
@@ -220,7 +250,7 @@ test_that("Validate selection short name - search_site", {
 
   expect_identical(
     search_site("toad", "posit", rlang = FALSE),
-    "https://community.posit.co/search?q=toad"
+    "https://forum.posit.co/search?q=toad"
   )
 
   expect_identical(